split up and rename test files
