+++ sjfnw/fund/tests/test_steps.py
@@ -1,3 +1,6 @@
+import logging
+import unittest
+
 from django.core.urlresolvers import reverse
 from django.test.utils import override_settings
 
@@ -5,129 +8,10 @@ from sjfnw.constants import TEST_MIDDLEWARE
 from sjfnw.fund import models
 from sjfnw.fund.tests.base import BaseFundTestCase, TEST_FIXTURE
 
-import logging, json, unittest
 logger = logging.getLogger('sjfnw')
 
-@override_settings(MIDDLEWARE_CLASSES = TEST_MIDDLEWARE,
-    PASSWORD_HASHERS = ('django.contrib.auth.hashers.MD5PasswordHasher',))
-class Home(BaseFundTestCase):
-
-  url = reverse('sjfnw.fund.views.home')
-
-  fixtures = TEST_FIXTURE
-
-  def setUp(self):
-    super(Home, self).setUp()
-    self.use_new_acct()
-
-  def test_new(self):
-    """ Verify that add mult form is shown to new memberships
-
-    Setup:
-      Member has no prior contacts
-      Login to membership in pre-training with 0 contacts
-      Login to membership in post-training with 0 contacts
-
-    Asserts:
-      Both use add_mult_flex.html
-      'likelihood' is in in the form for Post only
-    """
-
-    membership = models.Membership.objects.get(pk=self.pre_id)
-
-    response = self.client.get(self.url, follow=True)
-    self.assertTemplateUsed(response, 'fund/add_mult_flex.html')
-    self.assertEqual(response.context['request'].membership, membership)
-    self.assertNotContains(response, 'likelihood')
-
-    member = models.Member.objects.get(pk=self.member_id)
-    member.current = self.post_id
-    member.save()
-
-    response = self.client.get(self.url, follow=True)
-
-    membership = models.Membership.objects.get(pk=self.post_id)
-    self.assertTemplateUsed(response, 'fund/add_mult_flex.html')
-    self.assertEqual(response.context['request'].membership, membership)
-    self.assertContains(response, 'likelihood')
-
-  def test_contacts_without_est(self):
-
-    """ 2 contacts w/o est
-        logs into post training, gets estimates form
-        logs into pre, does not """
-
-    membership = models.Membership.objects.get(pk=self.pre_id)
-
-    contact1 = models.Donor(firstname='Anna', membership=membership)
-    contact1.save()
-    contact2 = models.Donor(firstname='Banana', membership=membership)
-    contact2.save()
-
-    response = self.client.get(self.url)
-    self.assertEqual(response.context['request'].membership, membership)
-    self.assertTemplateNotUsed('fund/add_estimates.html')
-
-    member = membership.member
-    member.current = self.post_id
-    member.save()
-
-    membership = models.Membership.objects.get(pk=self.post_id)
-
-    contact1.membership = membership
-    contact1.save()
-    contact2.membership = membership
-    contact2.save()
-
-    response = self.client.get(self.url, follow=True)
-    self.assertTemplateUsed('fund/add_estimates.html')
-    self.assertEqual(response.context['request'].membership, membership)
-
-  def test_estimates(self):
-
-    """ 2 contacts w/est
-        logs into post training, gets reg list """
-
-    membership = models.Membership.objects.get(pk=self.pre_id)
-
-    contact = models.Donor(firstname='Anna', membership=membership, amount=0, likelihood=0)
-    contact.save()
-    contact = models.Donor(firstname='Banana', membership=membership, amount=567, likelihood=34)
-    contact.save()
-
-    response = self.client.get(self.url)
-    self.assertTemplateNotUsed('fund/add_estimates.html')
-    self.assertEqual(response.context['request'].membership, membership)
-
-  @unittest.skip('Incomplete')
-  def test_contacts_list(self):
-    """ Verify correct display of a long contact list with steps, history
-
-    Setup:
-      Use membership 96 (test & gp 10) which has 29 contacts
-
-    Asserts:
-      ASSERTIONS
-    """
-
-    self.log_in_testy()
-    member = models.Member.objects.get(pk=1)
-    member.current = 96
-    member.save()
-
-    response = self.client.get(self.url)
-
-  @unittest.skip('Incomplete')
-  def test_gift_notification(self):
-    pass
-
-    """ add a gift to donor
-        test that notif shows up on next load
-        test that notif is gone on next load """
-
-
-@override_settings(MIDDLEWARE_CLASSES = TEST_MIDDLEWARE,
-    PASSWORD_HASHERS = ('django.contrib.auth.hashers.MD5PasswordHasher',))
+@override_settings(MIDDLEWARE_CLASSES=TEST_MIDDLEWARE,
+    PASSWORD_HASHERS=('django.contrib.auth.hashers.MD5PasswordHasher',))
 class StepComplete(BaseFundTestCase):
   """ Tests various scenarios of step completion
   1) completion
@@ -148,8 +32,9 @@ class StepComplete(BaseFundTestCase):
     super(StepComplete, self).setUp()
     logger.info('post super')
     self.use_test_acct()
-    self.url = reverse('sjfnw.fund.views.done_step',
-        kwargs = {'donor_id': self.donor_id, 'step_id': self.step_id})
+    self.url = reverse('sjfnw.fund.views.done_step', kwargs={
+        'donor_id': self.donor_id, 'step_id': self.step_id
+      })
     self.form_data = {
         'asked': 'on',
         'response': 2,
@@ -270,7 +155,6 @@ class StepComplete(BaseFundTestCase):
     self.assertEqual(response.content, "success")
 
     promised = form_data['promised_amount']
-    print(promised)
     if promised == '5,000': #hacky workaround
       promised = 5000
